refactor: extract processCrResultFiles and CrResultDeps into cr-pipeline.ts
Moves the CR analysis result processing logic out of the 1081-LOC
process-requests.ts into a dedicated cr-pipeline.ts module. Re-exports
the symbols from process-requests.ts for backwards compatibility.
Updates process-requests.test.ts to import directly from cr-pipeline.js.


Aloop-Agent: build
Aloop-Iteration: 4
Aloop-Session: orchestrator-20260321-172932-issue-169-20260324-085119
